requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.3.2
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
